+++ backend/src/main/java/com/digginroom/digginroom/oauth/OAuthIdTokenResolver.java
@@ -1,0 +1,5 @@
+package com.digginroom.digginroom.oauth;
+
+public interface OAuthIdTokenResolver {
+    String resolve(String idToken);
+}
