--- backend/src/main/java/com/digginroom/digginroom/service/OAuthUsernameResolver.java
@@ -1,5 +1,0 @@
-package com.digginroom.digginroom.service;
-
-public interface OAuthUsernameResolver {
-    String resolve(String idToken);
-}
